Created New Page for a new project
in this commit, I created a new page that is for a project I am working
on. Although most of the text is still placeholder text, I did change
the aside content and the brief description that is yet to be completed.
Also created a new folder that will be for projects. The project.html
file in said folder is to become the new default project
file/template.